app/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
gunicorn==23.0.0
pydantic[email]
email_validator==2.2.0
pydantic-settings==2.6.1
ujson==5.10.0
yarl==1.18.3
SQLAlchemy==2.0.36
opentelemetry-api==1.28.2
loguru==0.7.0
psycopg2-binary==2.9.10
python-jose>=3.0.0
pytz
Unidecode==1.3.8
requests==2.32.3
passlib==1.7.4
httptools==0.6.0
alembic==1.14.0
asyncpg==0.30.0
pyarmor
Pillow==10.4.0